refactor: moving current login related code into separate package
