+++ app/src/androidTest/java/uk/gov/android/authentication/login/LoginSessionConfigurationUtilTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.android.authentication
+package uk.gov.android.authentication.login
 
 import android.net.Uri
 import net.openid.appauth.AuthorizationRequest
