+++ app/src/main/java/uk/gov/android/authentication/login/TokenResponse.kt
@@ -1,7 +1,9 @@
-package uk.gov.android.authentication
+package uk.gov.android.authentication.login
 
+import kotlinx.serialization.InternalSerializationApi
 import kotlinx.serialization.Serializable
 
+@OptIn(InternalSerializationApi::class)
 @Serializable
 data class TokenResponse(
     val tokenType: String,
