+++ app/src/test/java/uk/gov/android/authentication/login/login/TokenResponseTest.kt
@@ -1,9 +1,10 @@
-package uk.gov.android.authentication
+package uk.gov.android.authentication.login.login
 
 import kotlinx.serialization.SerializationException
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import org.junit.jupiter.api.assertThrows
+import uk.gov.android.authentication.login.TokenResponse
 import kotlin.test.Test
 import kotlin.test.assertEquals
